requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12
tqdm>=4.65
pyyaml>=6.0
tensorboard>=2.14
scikit-learn>=1.3
